+++ kickr-server/src/main/java/com/kickr_server/dto/UserMatch/UserMatchUpdateDto.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class UserMatchUpdateDto {
-    int note;
+    double note;
     String comment;
